requirements-api.txt
# Requirements pour le hosted agent API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
openai>=1.12.0
python-dotenv>=1.0.0
